Shifted core logic and weather display to its own card
Since we want to generate multiples of the cards with different cities
